+++ src/termout.c
@@ -19,7 +19,7 @@
 
 #include <termios.h>
 #include <sys/time.h>
-#if CYGWIN_VERSION_API_MINOR >= 74
+#if CYGWIN_VERSION_API_MINOR >= 66
 #include <langinfo.h>  // nl_langinfo, CODESET
 #endif
 
@@ -1023,6 +1023,15 @@ write_char(wchar c, int width)
   last_char = c;
   last_attr = curs->attr;
 
+  wchar last_comb(termline *line, int col)
+  {
+    while (line->chars[col].cc_next) {
+      col += line->chars[col].cc_next;
+      //printf(" comb %04X\n", line->chars[col].chr);
+    }
+    return line->chars[col].chr;
+  }
+
   void put_char(wchar c)
   {
     if (term.ring_enabled && curs->x == term.marg_right + 1 - 8) {
@@ -1132,7 +1141,9 @@ write_char(wchar c, int width)
     //printf("ini %d:%d prev :%d\n", curs->y, curs->x, x);
     // if it's a pending emoji joined sequence, enforce handling of 
     // current character like a combining character
-    if (line->chars[x].attr.attr & TATTR_EMOJI)
+
+    //if (line->chars[x].attr.attr & TATTR_EMOJI)
+    if (last_comb(line, x) == 0x200D)
       //printf("@:%d (%04X) %04X prev joiner\n", x, line->chars[x].chr, c),
       width = 0;
   }
@@ -1204,6 +1215,8 @@ write_char(wchar c, int width)
       printf("write_char %04X %2d %08llX\n", c, width, curs->attr.attr);
 #endif
       if (curs->x > 0 || overstrike) {
+        bool fix_glitch = false;
+
        /* If we're in wrapnext state, the character
         * to combine with is _here_, not to our left. */
         int x = curs->x - !curs->wrapnext;
@@ -1213,12 +1226,32 @@ write_char(wchar c, int width)
        /*
         * If the previous character is UCSWIDE, back up another one.
         */
-        bool is_wide = false;
+        bool base_is_wide = false;
         if (line->chars[x].chr == UCSWIDE) {
           assert(x > 0);
           x--;
-          is_wide = true;
+          base_is_wide = true;
         }
+        else
+       /*
+        * Special case handling to avoid glitch:
+          in non-autowrap mode, in case a wide character at the end of line 
+          (where the cursor is stuck in VT100 stall-at-end-of-line position) 
+          is narrowed by VS15, we need to adjust position and set some flags.
+        */
+        if (!term.autowrap && line->chars[curs->x].chr == UCSWIDE) {
+          if (term.emoji_width && curs->x == term.cols - 1) {
+            curs->x --;
+            x = curs->x;
+            base_is_wide = true;
+            // flag to adjust cursor position below
+            fix_glitch = true;
+          }
+        }
+#ifdef debug_combining_position_adjustment
+        if (x >= term.cols - 2)
+          printf("[%d:%d] (x %d) %04X (base_wide %d) [x∓1] %04X %04X %04X lwrp %X cwrpnxt %d\n", curs->y, curs->x, x, c, base_is_wide, line->chars[x - 1].chr, line->chars[x].chr, line->chars[x + 1].chr, (line->lattr & (LATTR_WRAPPED | LATTR_WRAPPED2)), curs->wrapnext);
+#endif
         //printf("cur %d:%d prev :%d\n", curs->y, curs->x, x);
 
         if (term.emoji_width) {
@@ -1227,17 +1260,17 @@ write_char(wchar c, int width)
             before we add its low surrogate (add_cc below)
          */
           bool emoji_joiner = c == 0x200D && could_be_emoji_base(&line->chars[x]);
+#if 0
+#warning flaky approach
+          // abusing TATTR_EMOJI to mark a pending joiner was a bad idea
+          // it caused glitches like on U+26A0 U+200D‍ at a line end
           if (emoji_joiner)
             //printf("%d:%d (%04X) %04X mark joiner\n", curs->y, curs->x, line->chars[x].chr, c),
             line->chars[x].attr.attr |= TATTR_EMOJI;
           else
             line->chars[x].attr.attr &= ~TATTR_EMOJI;
+#endif
 
-          wchar last_comb(termline *line, int col) {
-            while (line->chars[col].cc_next)
-              col += line->chars[col].cc_next;
-            return line->chars[col].chr;
-          }
           bool is_fitzpatrick = false;
 
          /* Tune Fitzpatrick colour on non-emojis */
@@ -1259,7 +1292,7 @@ write_char(wchar c, int width)
           }
 
          /* Enforce wide with certain modifiers */
-          if (!is_wide &&
+          if (!base_is_wide &&
               // enforce emoji sequence on:
               // U+FE0F VARIATION SELECTOR-16
               // U+200D ZERO WIDTH JOINER
@@ -1298,6 +1331,33 @@ write_char(wchar c, int width)
               curs->x++;
             }
           }
+         /* Enforce narrow with VS15 */
+          else if (base_is_wide && c == 0xFE0E && term.emoji_width) {
+            // position x was set to the character to be combined with
+
+            // clear right half of previously wide character
+            // also prevent insert_char/term_check_boundary 
+            // from seeing a wide character and spoiling it
+            line->chars[x + 1] = basic_erase_char;
+
+            if (curs->wrapnext)
+              curs->wrapnext = false;
+            else if (term.autowrap || curs->x < term.cols)
+              // backspace cursor to right half of previously wide character
+              curs->x -= 1;
+#ifdef VS15_pull_back
+            if (cfg.VS15_pull_back) {
+              // delete wide right half and move up rest of line
+              // (kitty does not do this)
+              insert_char(-1);  // delete UCSWIDE
+              // if the line is LATTR_WRAPPED, should we pull back 
+              // subsequent lines too???
+            }
+#endif
+          }
+          // tweak special case fixed above
+          if (fix_glitch)
+            curs->x += 2;
         }
 
        /* Try to precompose with the previous cell's base codepoint;
@@ -2805,6 +2865,18 @@ get_mode(bool privatemode, int arg)
         return 2 - term.vt220_keys;
       when 2004:       /* xterm bracketed paste mode */
         return 2 - term.bracketed_paste;
+      when 1020: {
+#if CYGWIN_VERSION_API_MINOR >= 66
+        bool utf8 = strcmp(nl_langinfo(CODESET), "UTF-8") == 0;
+#else
+        bool utf8 = strstr(cs_get_locale(), ".65001");
+#endif
+        return 2 - utf8;
+      }
+      when 1021:
+        return 2 - cs_ambig_wide;
+      when 1022:
+        return 2 - term.emoji_width;
 
       /* Mintty private modes */
       when 7700:       /* CJK ambiguous width reporting */
@@ -4688,7 +4760,7 @@ respond_capabilities(void)
      )
     p += sprintf(p, "M");  // MOUSE
   p += sprintf(p, "Sc7");  // DECSCUSR
-#if CYGWIN_VERSION_API_MINOR >= 74
+#if CYGWIN_VERSION_API_MINOR >= 66
   if (0 == strcmp(nl_langinfo(CODESET), "UTF-8")) {
     p += sprintf(p, "U");  // UNICODE_BASIC
     p += sprintf(p, "Uw%d", UNICODE_VERSION / 100);  // UNICODE_WIDTHS